refactor(dashboard): extract mention utils SSOT from composer-v2 and quick-intervene
OnlineKeeper, MentionCandidate interfaces and 6 helper functions
(keeperNameFromTarget, mentionQueryFromMessage, trailingMentionNameFromMessage,
onlineKeeperNameForMention, mentionCandidates, replaceTrailingMentionDraft)
were duplicated across board/composer-v2.ts and ops/quick-intervene.ts.
Extract to lib/mention-utils.ts. Net -35 lines.
