+++ dashboard/src/components/ops/quick-intervene.ts
@@ -25,23 +25,21 @@ import {
 } from './ops-state'
 import { executeAction } from './helpers'
 import { isKeeperOffline } from '../../lib/keeper-predicates'
+import {
+  type OnlineKeeper,
+  keeperNameFromTarget,
+  mentionQueryFromMessage,
+  trailingMentionNameFromMessage,
+  onlineKeeperNameForMention,
+  mentionCandidates,
+  replaceTrailingMentionDraft,
+} from '../../lib/mention-utils'
 
 interface ComposerTarget {
   action_type: 'broadcast' | 'keeper_message'
   target_type: 'root' | 'keeper'
   target_id?: string
   label: string
-}
-
-interface OnlineKeeper {
-  name: string
-  status?: string
-}
-
-interface MentionCandidate {
-  name: string
-  status?: string
-  selected: boolean
 }
 
 const MODE_OPTIONS: Array<{ value: QuickComposerMode, label: string, description: string }> = [
@@ -51,49 +49,6 @@ const MODE_OPTIONS: Array<{ value: QuickComposerMode, label: string, description
 ]
 
 const MENTION_LISTBOX_ID = 'quick-intervene-mention-listbox'
-
-function keeperNameFromTarget(value: string): string | null {
-  if (!value.startsWith('keeper:')) return null
-  const name = value.slice('keeper:'.length).trim()
-  return name || null
-}
-
-function mentionQueryFromMessage(message: string): string | null {
-  const match = message.match(/(?:^|\s)@([A-Za-z0-9_.-]*)$/)
-  return match?.[1] ?? null
-}
-
-function trailingMentionNameFromMessage(message: string): string | null {
-  const match = message.match(/(?:^|\s)@([A-Za-z0-9_.-]+)\s*$/)
-  return match?.[1] ?? null
-}
-
-function onlineKeeperNameForMention(onlineKeepers: OnlineKeeper[], mentionName: string | null): string | null {
-  if (!mentionName) return null
-  const normalized = mentionName.toLowerCase()
-  return onlineKeepers.find(keeper => keeper.name.toLowerCase() === normalized)?.name ?? null
-}
-
-function mentionCandidates(onlineKeepers: OnlineKeeper[], query: string | null, selectedKeeper: string | null): MentionCandidate[] {
-  const normalizedQuery = query?.toLowerCase() ?? ''
-  return onlineKeepers
-    .filter(keeper => normalizedQuery === '' || keeper.name.toLowerCase().includes(normalizedQuery))
-    .map(keeper => ({
-      name: keeper.name,
-      status: keeper.status,
-      selected: keeper.name === selectedKeeper,
-    }))
-    .sort((a, b) => Number(b.selected) - Number(a.selected) || a.name.localeCompare(b.name))
-    .slice(0, 5)
-}
-
-function replaceTrailingMentionDraft(message: string, keeperName: string): string {
-  if (/(?:^|\s)@[A-Za-z0-9_.-]*$/.test(message)) {
-    return message.replace(/(^|\s)@[A-Za-z0-9_.-]*$/, `$1@${keeperName} `)
-  }
-  const spacer = message.trimEnd().length > 0 ? ' ' : ''
-  return `${message.trimEnd()}${spacer}@${keeperName} `
-}
 
 function chooseMentionTarget(keeperName: string): void {
   quickTarget.value = `keeper:${keeperName}`
